pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "A microservice boilerplate"
authors = [
    {name = "Nicolas Oliver",email = "nicolasoliver03@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "PyYAML (>=6.0,<7.0)",
]

[tool.poetry]

[tool.poetry.scripts]
lint = "scripts:pylint"
scan = "scripts:bandit"
unittest = "scripts:unittest"
app = "scripts:app"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mdformat = "^0.7.22"
pre-commit = "^4.2.0"
pylint = "^3.3.7"
bandit = "^1.8.6"
isort = "^6.0.1"
requests = "^2.32.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
